tests: OSX: cast st.Blksize to int64
On Linux, Blksize is already int64. But on OSX
it is int32 and that caused a compile failure.

https://github.com/rfjakob/gocryptfs/issues/15

+++ tests/test_helpers/helpers.go
@@ -298,5 +298,6 @@ func Du(t *testing.T, fd int) (nBytes int64, nBlocks int64) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	return st.Blocks * st.Blksize, st.Blocks
+	// On OSX, Blksize is int32, need to cast to int64.
+	return st.Blocks * int64(st.Blksize), st.Blocks
 }
